LanGraph/multi_agent_app/requirements.txt
streamlit
langgraph
langchain
langchain-community
langchain-ollama
langchain-text-splitters
chromadb
sentence-transformers
pyautogen
openai